jiva-health/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
firebase-admin==6.4.0
google-cloud-vision==3.5.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
spacy==3.7.2
nltk==3.8.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pillow==10.2.0
aiofiles==23.2.1
httpx==0.26.0
